Revert "feat: Addes MetaInfoPanel and descriptions for the ComboBox entries"
