+++ query-by-graph/build
@@ -1,4 +1,4 @@
 #!/bin/sh
 
 wasm-pack build
-npm run dev
+npm run build
