+++ query-by-graph/package.json
@@ -9,11 +9,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@guolao/vue-monaco-editor": "^1.5.4",
     "@headlessui/vue": "^1.7.22",
     "axios": "^1.7.4",
     "highlight.js": "^11.10.0",
-    "monaco-editor": "^0.52.0",
     "rete": "^2.0.3",
     "rete-area-plugin": "^2.0.4",
     "rete-connection-path-plugin": "^2.0.1",
