+++ query-by-graph/src/lib/rete/editor.ts
@@ -185,9 +185,6 @@ export async function createEditor(container: HTMLElement) {
 
     connection.addPreset(ConnectionPresets.classic.setup());
 
-
-    // this is called when something in the editor happens.
-    // you can get a list of event keywords here: https://rete.readthedocs.io/en/latest/Events/
     area.addPipe(async (context) => {
         // this is a workaround to hinder the counter from increasing at every
         // draw method of the editor
@@ -281,18 +278,6 @@ export async function createEditor(container: HTMLElement) {
                     await editor.removeConnection(c.id);
                 }
                 await editor.removeNode(source.id);
-            }
-        }
-
-        if(context.type === 'nodepicked') {
-            const node = context.data as ClassicPreset.Node;
-            console.log(`Node clicked: ${node.id}`);
-
-            if(vueCallback){
-                vueCallback({
-                    type: 'nodeselected',
-                    data: node,
-                });
             }
         }
