+++ query-by-graph/tailwind.config.js
@@ -1,7 +1,4 @@
-// require ist in ESModul nicht erlaubt, deshalb Import
 /** @type {import('tailwindcss').Config} */
-import forms from '@tailwindcss/forms';
-
 export default {
   content: [
     "./index.html",
@@ -11,6 +8,6 @@ export default {
     extend: {},
   },
   plugins: [
-    forms, // Statt require() direkt als Import
+      require("@tailwindcss/forms")
   ],
-};
+}
